+++ RELEASE_NOTES.md
@@ -9,6 +9,7 @@ will also work as expected.)* *(Issue #61)*
 * Fix an issue with combining `options.filter` and `options.includeInvalid` *(Issue #63)*
 * We now clone the JSON Reference definition and JSON Reference details respectively for `options.refPreProcessor' and
 `options.refPostProcessor` *(Issue #64)*
+* Fix an issue where a JSON Reference at the root of the document was not resolved properly *(Issue #65)*
 
 ### v2.0.4 (2016-01-21)